changed file to js and deleted compiled js file.
