+++ src/utils/analytics.js
@@ -1,80 +1,41 @@
-"use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (g && (g = 0, op[0] && (_ = 0)), _) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+import {
+    Identify,
+    identify,
+    init,
+    reset,
+    setUserId,
+    track,
+  } from "@amplitude/analytics-browser";
+  
+  class AmplitudeAnalytics {
+    initialise() {
+      init(process.env.REACT_APP_AMPLITUDE_KEY || "", {
+        defaultTracking: true,
+      });
     }
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-var analytics_browser_1 = require("@amplitude/analytics-browser");
-var AmplitudeAnalytics = /** @class */ (function () {
-    function AmplitudeAnalytics() {
+  
+    identifyUser(userId) {
+      setUserId(userId);
     }
-    AmplitudeAnalytics.prototype.initialise = function () {
-        (0, analytics_browser_1.init)(process.env.REACT_APP_AMPLITUDE_KEY || "", {
-            defaultTracking: true,
-        });
-    };
-    AmplitudeAnalytics.prototype.identifyUser = function (userId) {
-        (0, analytics_browser_1.setUserId)(userId);
-    };
-    AmplitudeAnalytics.prototype.identifyUserProperties = function (properties) {
-        var id = new analytics_browser_1.Identify();
-        var params = Object.entries(properties);
-        for (var _i = 0, params_1 = params; _i < params_1.length; _i++) {
-            var _a = params_1[_i], key = _a[0], value = _a[1];
-            // id.set(param, params[param]);
-            // console.log(key, value);
-            id.set(key, value);
-        }
-        (0, analytics_browser_1.identify)(id);
-    };
-    AmplitudeAnalytics.prototype.trackEvent = function (eventName, eventProps) {
-        (0, analytics_browser_1.track)(eventName, eventProps);
-    };
-    AmplitudeAnalytics.prototype.resetAmplitude = function () {
-        return __awaiter(this, void 0, void 0, function () {
-            return __generator(this, function (_a) {
-                switch (_a.label) {
-                    case 0: return [4 /*yield*/, (0, analytics_browser_1.track)("Logout")];
-                    case 1:
-                        _a.sent();
-                        (0, analytics_browser_1.reset)();
-                        return [2 /*return*/];
-                }
-            });
-        });
-    };
-    return AmplitudeAnalytics;
-}());
-exports.default = new AmplitudeAnalytics();
+  
+    setUserParams(params) {
+      const user = new Identify();
+  
+      for (const param of params) {
+        user.set(param.name, param.value)
+      }
+  
+      identify(user)
+    }
+  
+    trackEvent(eventName, params = null) {
+      track(`LandingPage: ${eventName}`, params);
+    }
+  
+    async resetAmplitude() {
+      await track("Logout").promise();
+      reset();
+    }
+  }
+  
+  export default new AmplitudeAnalytics();
